+++ components/sql/src/main/scala/pl/touk/nussknacker/sql/DatabaseEnricherComponentProvider.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.sql
 
 import com.typesafe.config.Config
 import net.ceedubs.ficus.Ficus._
-import net.ceedubs.ficus.readers.ArbitraryTypeReader._
 import pl.touk.nussknacker.engine.api.component.{
   ComponentDefinition,
   ComponentDependencies,
@@ -10,7 +9,6 @@ import pl.touk.nussknacker.engine.api.component.{
   NussknackerVersion
 }
 import pl.touk.nussknacker.engine.util.config.DocsConfig
-import pl.touk.nussknacker.sql.db.schema.MetaDataProviderFactory
 import pl.touk.nussknacker.sql.service.{DatabaseLookupEnricher, DatabaseQueryEnricher}
 
 class DatabaseEnricherComponentProvider extends ComponentProvider {
@@ -36,23 +34,22 @@ object DatabaseEnricherComponentProvider {
     val docsConfig = DocsConfig(config)
     import docsConfig._
 
-    val factory         = new MetaDataProviderFactory()
     val componentConfig = config.getConfig("config")
     val dbQueryEnricher = readEnricherConfigIfPresent(componentConfig, "databaseQueryEnricher")
       .map(queryConfig =>
         ComponentDefinition(
-          name = queryConfig.name,
+          name = queryConfig.name.value,
           component = new DatabaseQueryEnricher(
             queryConfig.dbPool,
-            factory.create(queryConfig.dbPool)
+            queryConfig.name
           )
         ).withRelativeDocs("Enrichers#databasequeryenricher")
       )
     val dbLookupEnricher = readEnricherConfigIfPresent(componentConfig, "databaseLookupEnricher")
       .map(lookupConfig =>
         ComponentDefinition(
-          name = lookupConfig.name,
-          component = new DatabaseLookupEnricher(lookupConfig.dbPool, factory.create(lookupConfig.dbPool))
+          name = lookupConfig.name.value,
+          component = new DatabaseLookupEnricher(lookupConfig.dbPool, lookupConfig.name)
         ).withRelativeDocs("Enrichers#databaselookupenricher")
       )
 
@@ -60,11 +57,12 @@ object DatabaseEnricherComponentProvider {
 
   }
 
-  private def readEnricherConfigIfPresent(config: Config, path: String): Option[DbEnricherConfig] =
+  private def readEnricherConfigIfPresent(config: Config, path: String): Option[DbEnricherConfig] = {
     if (config.hasPath(path)) {
       Some(config.as[DbEnricherConfig](path))
     } else {
       None
     }
+  }
 
 }