+++ components/sql/src/main/scala/pl/touk/nussknacker/sql/db/pool/DBPoolConfig.scala
@@ -13,4 +13,5 @@ final case class DBPoolConfig(
     timeout: Duration = FiniteDuration(30, TimeUnit.SECONDS),
     schema: Option[String] = None,
     dataSourceProperties: Map[String, String] = Map.empty,
+    registerMbeans: Boolean = true
 )