+++ components/sql/src/main/scala/pl/touk/nussknacker/sql/db/pool/HikariDataSourceFactory.scala
@@ -4,8 +4,9 @@ import com.zaxxer.hikari.{HikariConfig, HikariDataSource}
 
 object HikariDataSourceFactory {
 
-  def apply(conf: DBPoolConfig): HikariDataSource = {
+  def apply(conf: DBPoolConfig, poolName: String): HikariDataSource = {
     val hikariConf = new HikariConfig()
+    hikariConf.setPoolName(poolName)
     hikariConf.setJdbcUrl(conf.url)
     hikariConf.setUsername(conf.username)
     hikariConf.setPassword(conf.password)
@@ -17,6 +18,7 @@ object HikariDataSourceFactory {
     conf.dataSourceProperties.foreach { case (name, value) =>
       hikariConf.addDataSourceProperty(name, value)
     }
+    hikariConf.setRegisterMbeans(conf.registerMbeans)
     new HikariDataSource(hikariConf)
   }