+++ docs/Changelog.md
@@ -302,6 +302,7 @@ description: Stay informed with detailed changelogs covering new features, impro
 * [#8606](https://github.com/TouK/nussknacker/pull/8606) `Testkit`: Added `TestNodeCompiler`
 * [#8733](https://github.com/TouK/nussknacker/pull/8733) Added support for Basic Authentication in OpenAPI Enricher
 * [#8781](https://github.com/TouK/nussknacker/pull/8781) Removed `WithExceptionHandler` trait, it made it easy to unintentionally hide original `close` method
+* [#8802](https://github.com/TouK/nussknacker/pull/8802) SQL enrichers: sql metadata uses connection pooling
 
 ## 1.18